labs/lab4-markdown-injection/server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
openai>=1.31.0
python-dotenv>=1.0.0
Pillow>=10.4.0
